added timer

+++ src/main/java/org/okcrobot/scouter/model/timer/TimerListener.java
@@ -2,6 +2,6 @@ package org.okcrobot.scouter.model.timer;
 
 public interface TimerListener {
   
-  public void update(int time);
+  public void update(long time);
   
 }
